requirements.txt
python-dotenv>=1.0.0
mysql-connector-python>=8.0.0
discord.py>=2.6.0
yt-dlp>=2024.1.0
psutil>=5.9.0
Pillow>=9.0.0
rembg[cpu]>=2.0.0
opencv-python-headless>=4.5.0
numpy>=1.20.0
